fix: [demo] better color struct syntax

+++ demo/Code/Scenes/Gameplay.hh
@@ -5,7 +5,6 @@ struct Gameplay final : gph::Scene {
     // NOTE: the mouse isn't confined to the window yet, and hidding it makes it
     //       complicated to control when moved outside the window (remains hidden).
     // cbn::win::SetMouseVisibility(false);
-
     auto s_ch = r_AssetMgr.LoadSprite("Art/Sprites/crosshair.png");
     {// Entities
       auto &e_ch = NewEntity<Crosshair, gph::RenderLayer::UI>(s_ch);
@@ -19,6 +18,6 @@ struct Gameplay final : gph::Scene {
   }
 
   virtual void Render(void) const override {
-    r_Canvas.Fill((u32)Color::Black);
+    r_Canvas.Fill(Color::Black);
   }
 };
